backend/requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
openai==2.3.0
python-dotenv==1.0.1
requests==2.31.0

# ChromaDB dependencies
chromadb==0.4.18
sentence-transformers>=3.0.0

# Additional utilities
numpy>=1.26.0,<2.0.0
scipy>=1.13.0

# File processing libraries for knowledge base upload
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0